Moves Team types to Prytaneum namespace
