--- src/components/Team/types.tsx
@@ -1,15 +1,0 @@
-interface reference {
-    link: string;
-    name: string;
-    icon: JSX.Element;
-}
-
-export interface teamMember {
-    picturePath?: string;
-    fullName: string;
-    subtitle: string;
-    description: string;
-    startDate: Date;
-    endDate: Date;
-    references?: reference[];
-}
